+++ packages/orq-rc/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [typescript v1.41.0-rc.8] packages/orq-rc
 ### Releases
 - [NPM v1.41.0-rc.8] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.41.0-rc.8 - packages/orq-rc
+
+## 2024-11-24 15:49:34
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.444.3 (2.461.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.41.0-rc.15] packages/orq-rc
+### Releases
+- [NPM v1.41.0-rc.15] https://www.npmjs.com/package/orq-poc-typescript-multi-env-version/v/1.41.0-rc.15 - packages/orq-rc